1.Added leaderboard page 2. Handled post question submission 3. Fixed bugs

+++ db/migrations/20260127201241_create_leaderboard.up.sql
@@ -3,8 +3,8 @@ CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
 CREATE TABLE LEADERBOARD (
     contest_id   UUID NOT NULL,
     user_id      UUID NOT NULL,
-    score        SMALLINT NOT NULL,
-    rank         SMALLINT NOT NULL,
+    score        INTEGER NOT NULL,
+    rank         INTEGER NOT NULL,
     created_at   BIGINT NOT NULL DEFAULT EXTRACT(EPOCH FROM NOW()),
     PRIMARY KEY  (contest_id, user_id)
 )
